+++ config/module.py
@@ -35,24 +35,20 @@ def loadModule():
 
     pic32cx_bz3_family = {'PIC32CX5109BZ31048',
                           'PIC32CX5109BZ31032',
-                          'WBZ435',
+                          'WBZ351',
+                          'WBZ350',
                           }
     processor = Variables.get('__PROCESSOR')
     print('processor={}'.format(processor))
 
+    if((processor in pic32cx_bz2_family) or (processor in pic32cx_bz3_family)):
+        ## Device Support
+        execfile(Module.getPath() + '/config/module_pic32cx_bz_device_support.py')
+        ## Persistant Data Storage
+        execfile(Module.getPath() + '/config/module_pic32cx_bz_pds.py')    
+    
     if( processor in pic32cx_bz2_family):
-        ## PIC32CX-BZ System_Service
-        execfile(Module.getPath() + '/config/module_pic32cx_bz2_device_support.py')
-        ## Persistant Data Storage
-        execfile(Module.getPath() + '/config/module_pic32cx_bz_pds.py')
-        ## PIC32CX-BZ Bootloader
+        ## PIC32CX-BZ2 Bootloader
         execfile(Module.getPath() + '/config/module_pic32cx_bz2_bootloader.py')
-        ## PIC32CX-BZ OTA
+        ## PIC32CX-BZ2 Bootloader services
         execfile(Module.getPath() + '/config/module_pic32cx_bz2_bootloaderServices.py')
-
-    elif( processor in pic32cx_bz3_family):
-        ## PIC32CX-BZ System_Service
-        execfile(Module.getPath() + '/config/module_pic32cx_bz3_device_support.py')
-        ## Persistant Data Storage
-        execfile(Module.getPath() + '/config/module_pic32cx_bz_pds.py')
-        execfile(Module.getPath() + '/config/module_pic32cx_bz3_signFW.py')