+++ driver/pic32cx-bz/config/device_support.py
@@ -35,10 +35,20 @@ pic32cx_bz2_family = {'PIC32CX1012BZ25048',
                       'WBZ450',
                       }
 
+pic32cx_bz2_48pin_family = {'PIC32CX1012BZ25048',
+                            'WBZ451',
+                            }
+
 pic32cx_bz3_family = {'PIC32CX5109BZ31048',
                       'PIC32CX5109BZ31032',
-                      'WBZ435',
+                      'WBZ351',
+                      'WBZ350',
                       }
+
+pic32cx_bz3_48pin_family = {'PIC32CX5109BZ31048',
+                            'WBZ351',
+                           }
+
 global deviceName
 deviceName = Variables.get("__PROCESSOR")
 
@@ -58,7 +68,9 @@ REQUIRES_APP_IDLE_TASK = {
         'ZIGBEE_CUSTOM': 'ZIGBEESTACK_LOADED',
         'ZIGBEE_MACWSNTESTER': 'ZIGBEESTACK_LOADED',
         'ZIGBEE_GPD_SENSOR': 'ZIGBEESTACK_LOADED',
-        'ZIGBEE_ZAPPSI': 'ZIGBEESTACK_LOADED'}
+        'ZIGBEE_ZAPPSI': 'ZIGBEESTACK_LOADED',
+        'OPEN_THREAD': 'THREADSTACK_LOADED',
+        'IEEE_802154_PHY': '802154_PHY_LOADED'}
 
 
 REQUIRES_RTC_SUPPORT = {}
@@ -80,6 +92,8 @@ RADIOSTACK_COMPONENTS = [
         'ZIGBEE_MACWSNTESTER',
         'ZIGBEE_GPD_SENSOR',
         'ZIGBEE_ZAPPSI',
+        'OPEN_THREAD',
+        'IEEE_802154_PHY'
         ]
 
 activeComponents = Database.getActiveComponentIDs()
@@ -125,6 +139,12 @@ def configWSSEnable(component, targetID, connected):
                         format(stackName, str(REQUIRES_WSSENABLE_MODE[stackName])))
 
                 if 'ZIGBEE' in stackName and REQUIRES_WSSENABLE_MODE[stackName]:
+                    zigbee_loaded = True
+                    
+                if 'IEEE802154_PHY' in stackName and REQUIRES_WSSENABLE_MODE[stackName]:
+                    zigbee_loaded = True
+
+                if 'OPEN_THREAD' in stackName and REQUIRES_WSSENABLE_MODE[stackName]:
                     zigbee_loaded = True
 
             if ble_loaded and zigbee_loaded:
@@ -298,6 +318,37 @@ def sendAntMessage(id, value):
             Log.writeInfoMessage('{:<17}: Sending  - target={} ID={} Value={}'.format('device_support', client, id, value))
             Database.sendMessage(client, "ANTENNA_GAIN_CHANGE", {'target': client, id:value})
 
+def getBLEStackLibDsadven():
+    components = Database.getActiveComponentIDs()
+    if('BLE_STACK_LIB' in components):
+        dsadvSetting = Database.getSymbolValue("BLE_STACK_LIB", "GAP_DSADV_EN")
+        return dsadvSetting
+    return False
+
+def getDeepSleepState():
+    if (deepSleepEnable.getValue()):
+        return True
+    return False
+
+def computeRetentionRAMRequirements():
+    global stackRequirements
+    print('computing Retention RAM Requirements')
+    stackRequirements = 0
+    components = Database.getActiveComponentIDs()
+    if('BLE_STACK_LIB' in components):
+        stackRequirements = 616
+    if any("ZIGBEE" in s for s in components):
+        stackRequirements = stackRequirements + 2048
+    return stackRequirements
+
+def computeMaxAppRetentionRAM():
+    appRequirements = 0
+    if (deviceName in pic32cx_bz2_family):
+        appRequirements = (8192 - stackRetentionRAM.getValue())
+    elif (deviceName in pic32cx_bz3_family):
+        appRequirements = (32768 - stackRetentionRAM.getValue())
+    return appRequirements
+
 def instantiateComponent(libBTZBCore):
     print('PIC32CX-BZ BTZB_Common')
     configName = Variables.get('__CONFIGURATION_NAME')
@@ -388,10 +439,10 @@ def instantiateComponent(libBTZBCore):
     sys_service_a = libBTZBCore.createLibrarySymbol(None, None)
     sys_service_a.setDestPath('driver/device_support')
     if (deviceName in pic32cx_bz2_family):
-        sys_service_a.setSourcePath('driver/pic32cx-bz/src/pic32cx_bz2_device_support.a')
+        sys_service_a.setSourcePath('driver/pic32cx-bz/src/src_bz2/device_support/pic32cx_bz2_device_support.a')
         sys_service_a.setOutputName('pic32cx_bz2_device_support.a')
     elif (deviceName in pic32cx_bz3_family):
-        sys_service_a.setSourcePath('driver/pic32cx-bz/src/pic32cx_bz3_device_support.a')
+        sys_service_a.setSourcePath('driver/pic32cx-bz/src/src_bz3/device_support/pic32cx_bz3_device_support.a')
         sys_service_a.setOutputName('pic32cx_bz3_device_support.a')
 
 
@@ -401,7 +452,10 @@ def instantiateComponent(libBTZBCore):
 
     # Add rf_system.h file
     rfSysInitializeHeaderFile = libBTZBCore.createFileSymbol(None, None)
-    rfSysInitializeHeaderFile.setSourcePath('driver/pic32cx-bz/src/rf_system.h')
+    if (deviceName in pic32cx_bz2_family):
+        rfSysInitializeHeaderFile.setSourcePath('driver/pic32cx-bz/src/src_bz2/device_support/rf_system.h')
+    elif (deviceName in pic32cx_bz3_family):
+        rfSysInitializeHeaderFile.setSourcePath('driver/pic32cx-bz/src/src_bz3/device_support/rf_system.h')
     rfSysInitializeHeaderFile.setOutputName('rf_system.h')
     rfSysInitializeHeaderFile.setOverwrite(True)
     rfSysInitializeHeaderFile.setDestPath('driver/device_support/include')
@@ -411,7 +465,10 @@ def instantiateComponent(libBTZBCore):
 
     # Add pmu_system.h file
     pmuSystemHeaderFile = libBTZBCore.createFileSymbol(None, None)
-    pmuSystemHeaderFile.setSourcePath('driver/pic32cx-bz/src/pmu_system.h')
+    if (deviceName in pic32cx_bz2_family):
+        pmuSystemHeaderFile.setSourcePath('driver/pic32cx-bz/src/src_bz2/device_support/pmu_system.h')
+    elif (deviceName in pic32cx_bz3_family):
+        pmuSystemHeaderFile.setSourcePath('driver/pic32cx-bz/src/src_bz3/device_support/pmu_system.h')
     pmuSystemHeaderFile.setOutputName('pmu_system.h')
     pmuSystemHeaderFile.setOverwrite(True)
     pmuSystemHeaderFile.setDestPath('driver/device_support/include')
@@ -421,7 +478,10 @@ def instantiateComponent(libBTZBCore):
     
     # Add info_block.h file
     infoBlockHeaderFile = libBTZBCore.createFileSymbol(None, None)
-    infoBlockHeaderFile.setSourcePath('driver/pic32cx-bz/src/info_block.h')
+    if (deviceName in pic32cx_bz2_family):
+        infoBlockHeaderFile.setSourcePath('driver/pic32cx-bz/src/src_bz2/device_support/info_block.h')
+    elif (deviceName in pic32cx_bz3_family):
+        infoBlockHeaderFile.setSourcePath('driver/pic32cx-bz/src/src_bz3/device_support/info_block.h')
     infoBlockHeaderFile.setOutputName('info_block.h')
     infoBlockHeaderFile.setOverwrite(True)
     infoBlockHeaderFile.setDestPath('driver/device_support/include')
@@ -431,7 +491,10 @@ def instantiateComponent(libBTZBCore):
     
     # Add sleep_system.h file
     sleepSystemHeaderFile = libBTZBCore.createFileSymbol(None, None)
-    sleepSystemHeaderFile.setSourcePath('driver/pic32cx-bz/src/sleep_system.h')
+    if (deviceName in pic32cx_bz2_family):
+        sleepSystemHeaderFile.setSourcePath('driver/pic32cx-bz/src/src_bz2/device_support/sleep_system.h')
+    elif (deviceName in pic32cx_bz3_family):
+        sleepSystemHeaderFile.setSourcePath('driver/pic32cx-bz/src/src_bz3/device_support/sleep_system.h')
     sleepSystemHeaderFile.setOutputName('sleep_system.h')
     sleepSystemHeaderFile.setOverwrite(True)
     sleepSystemHeaderFile.setDestPath('driver/device_support/include')
@@ -441,7 +504,10 @@ def instantiateComponent(libBTZBCore):
     
     # Add framework_defs.h
     HeaderFile = libBTZBCore.createFileSymbol(None, None)
-    HeaderFile.setSourcePath('driver/pic32cx-bz/src/framework_defs.h')
+    if (deviceName in pic32cx_bz2_family):
+        HeaderFile.setSourcePath('driver/pic32cx-bz/src/src_bz2/app_fw/framework_defs.h')
+    elif (deviceName in pic32cx_bz3_family):
+        HeaderFile.setSourcePath('driver/pic32cx-bz/src/src_bz3/app_fw/framework_defs.h')
     HeaderFile.setOutputName('framework_defs.h')
     HeaderFile.setOverwrite(True)
     HeaderFile.setDestPath('')
@@ -451,8 +517,37 @@ def instantiateComponent(libBTZBCore):
 
     # Add osal_freertos_extend.h
     HeaderFile = libBTZBCore.createFileSymbol(None, None)
-    HeaderFile.setSourcePath('driver/pic32cx-bz/src/osal/osal_freertos_extend.h')
+    if (deviceName in pic32cx_bz2_family):
+        HeaderFile.setSourcePath('driver/pic32cx-bz/src/src_bz2/app_fw/osal_freertos_extend.h')
+    elif (deviceName in pic32cx_bz3_family):
+        HeaderFile.setSourcePath('driver/pic32cx-bz/src/src_bz3/app_fw/osal_freertos_extend.h')
     HeaderFile.setOutputName('osal_freertos_extend.h')
+    HeaderFile.setOverwrite(True)
+    HeaderFile.setDestPath('/osal')
+    HeaderFile.setProjectPath('config/' + configName + '/osal/')
+    HeaderFile.setType('HEADER')
+    HeaderFile.setEnabled(True)
+
+    # Add osal_freertos.h
+    HeaderFile = libBTZBCore.createFileSymbol(None, None)
+    if (deviceName in pic32cx_bz2_family):
+        HeaderFile.setSourcePath('driver/pic32cx-bz/src/src_bz2/app_fw/osal_freertos.h')
+    elif (deviceName in pic32cx_bz3_family):
+        HeaderFile.setSourcePath('driver/pic32cx-bz/src/src_bz3/app_fw/osal_freertos.h')
+    HeaderFile.setOutputName('osal_freertos.h')
+    HeaderFile.setOverwrite(True)
+    HeaderFile.setDestPath('/osal')
+    HeaderFile.setProjectPath('config/' + configName + '/osal/')
+    HeaderFile.setType('HEADER')
+    HeaderFile.setEnabled(True)
+
+    # Add osal.h
+    HeaderFile = libBTZBCore.createFileSymbol(None, None)
+    if (deviceName in pic32cx_bz2_family):
+        HeaderFile.setSourcePath('driver/pic32cx-bz/src/src_bz2/app_fw/osal.h')
+    elif (deviceName in pic32cx_bz3_family):
+        HeaderFile.setSourcePath('driver/pic32cx-bz/src/src_bz3/app_fw/osal.h')
+    HeaderFile.setOutputName('osal.h')
     HeaderFile.setOverwrite(True)
     HeaderFile.setDestPath('/osal')
     HeaderFile.setProjectPath('config/' + configName + '/osal/')
@@ -466,8 +561,25 @@ def instantiateComponent(libBTZBCore):
 
     # Add osal_freertos_extend.c
     SourceFile = libBTZBCore.createFileSymbol(None, None)
-    SourceFile.setSourcePath('driver/pic32cx-bz/src/osal/osal_freertos_extend.c')
+    if (deviceName in pic32cx_bz2_family):
+        SourceFile.setSourcePath('driver/pic32cx-bz/src/src_bz2/app_fw/osal_freertos_extend.c')
+    elif (deviceName in pic32cx_bz3_family):
+        SourceFile.setSourcePath('driver/pic32cx-bz/src/src_bz3/app_fw/osal_freertos_extend.c')
     SourceFile.setOutputName('osal_freertos_extend.c')
+    SourceFile.setOverwrite(True)
+    SourceFile.setDestPath('/osal')
+    SourceFile.setProjectPath('config/' + configName + '/osal/')
+    SourceFile.setType('SOURCE')
+    SourceFile.setEnabled(True)
+
+   
+    # Add osal_freertos.c
+    SourceFile = libBTZBCore.createFileSymbol(None, None)
+    if (deviceName in pic32cx_bz2_family):
+        SourceFile.setSourcePath('driver/pic32cx-bz/src/src_bz2/app_fw/osal_freertos.c')
+    elif (deviceName in pic32cx_bz3_family):
+        SourceFile.setSourcePath('driver/pic32cx-bz/src/src_bz3/app_fw/osal_freertos.c')
+    SourceFile.setOutputName('osal_freertos.c')
     SourceFile.setOverwrite(True)
     SourceFile.setDestPath('/osal')
     SourceFile.setProjectPath('config/' + configName + '/osal/')
@@ -496,9 +608,15 @@ def instantiateComponent(libBTZBCore):
     HeaderFile.setMarkup(True)
     HeaderFile.setEnabled(True)
 
+    # Application definitions from BLE
+    dsAppBleMsgIdList = libBTZBCore.createListSymbol("LIST_DS_BLE_MSG_ID_H", None)
+
     # Add app_idle_task.h
     freertosidleTaskHeaderFile = libBTZBCore.createFileSymbol("APP_IDLE_TASK_H", None)
-    freertosidleTaskHeaderFile.setSourcePath("driver/pic32cx-bz/src/app_idle_task.h")
+    if (deviceName in pic32cx_bz2_family):
+        freertosidleTaskHeaderFile.setSourcePath("driver/pic32cx-bz/src/src_bz3/app_fw/app_idle_task.h")
+    elif (deviceName in pic32cx_bz3_family):
+        freertosidleTaskHeaderFile.setSourcePath("driver/pic32cx-bz/src/src_bz3/app_fw/app_idle_task.h")
     freertosidleTaskHeaderFile.setOutputName("app_idle_task.h")
     freertosidleTaskHeaderFile.setDestPath('../../')
     freertosidleTaskHeaderFile.setProjectPath('')
@@ -513,7 +631,7 @@ def instantiateComponent(libBTZBCore):
     ############################################################################
 
     # Add app.c
-    SourceFile = libBTZBCore.createFileSymbol(None, None)
+    SourceFile = libBTZBCore.createFileSymbol("DEVICE_APP_C", None)
     SourceFile.setSourcePath('driver/pic32cx-bz/templates/app.c.ftl')
     SourceFile.setOutputName('app.c')
     SourceFile.setOverwrite(True)
@@ -522,6 +640,12 @@ def instantiateComponent(libBTZBCore):
     SourceFile.setType('SOURCE')
     SourceFile.setEnabled(True)
     SourceFile.setMarkup(True)
+
+    # Application codes from BLE
+    dsAppBleIncludeList = libBTZBCore.createListSymbol("LIST_DS_BLE_INCLUDE_C", None)
+    dsAppBleDataList = libBTZBCore.createListSymbol("LIST_DS_BLE_DATA_C", None)
+    dsAppBleInitList = libBTZBCore.createListSymbol("LIST_DS_BLE_INIT_C", None)
+    dsAppBleTaskEntryList = libBTZBCore.createListSymbol("LIST_DS_BLE_TASK_ENTRY_C", None)
 
     # Add app_idle_task.c
     freertosidleTaskSourceFile = libBTZBCore.createFileSymbol("APP_IDLE_TASK_C", None)
@@ -548,7 +672,10 @@ def instantiateComponent(libBTZBCore):
     # Add device_deep_sleep.h
     global deviceDeepSleepHeaderFile
     deviceDeepSleepHeaderFile = libBTZBCore.createFileSymbol("DEVICE_DEEP_SLEEP_H", None)
-    deviceDeepSleepHeaderFile.setSourcePath("driver/pic32cx-bz/src/deep_sleep/device_deep_sleep.h")
+    if (deviceName in pic32cx_bz2_family):
+        deviceDeepSleepHeaderFile.setSourcePath("driver/pic32cx-bz/src/src_bz2/app_fw/device_deep_sleep.h")
+    elif (deviceName in pic32cx_bz3_family):
+        deviceDeepSleepHeaderFile.setSourcePath("driver/pic32cx-bz/src/src_bz3/app_fw/device_deep_sleep.h")
     deviceDeepSleepHeaderFile.setOutputName("device_deep_sleep.h")
     deviceDeepSleepHeaderFile.setDestPath('')
     deviceDeepSleepHeaderFile.setProjectPath('config/' + configName)
@@ -560,7 +687,8 @@ def instantiateComponent(libBTZBCore):
     # Add device_deep_sleep.c
     global deviceDeepSleepSourceFile
     deviceDeepSleepSourceFile = libBTZBCore.createFileSymbol("DEVICE_DEEP_SLEEP_C", None)
-    deviceDeepSleepSourceFile.setSourcePath("driver/pic32cx-bz/src/deep_sleep/device_deep_sleep.c")
+
+    deviceDeepSleepSourceFile.setSourcePath("driver/pic32cx-bz/templates/low_power/device_deep_sleep.ftl")
     deviceDeepSleepSourceFile.setOutputName("device_deep_sleep.c")
     deviceDeepSleepSourceFile.setDestPath('')
     deviceDeepSleepSourceFile.setProjectPath('config/' + configName)
@@ -573,7 +701,10 @@ def instantiateComponent(libBTZBCore):
     # Add device_sleep.h
     global deviceSleepHeaderFile
     deviceSleepHeaderFile = libBTZBCore.createFileSymbol("DEVICE_SLEEP_H", None)
-    deviceSleepHeaderFile.setSourcePath("driver/pic32cx-bz/src/sleep/device_sleep.h")
+    if (deviceName in pic32cx_bz2_family):
+        deviceSleepHeaderFile.setSourcePath("driver/pic32cx-bz/src/src_bz2/app_fw/device_sleep.h")
+    elif (deviceName in pic32cx_bz3_family):
+        deviceSleepHeaderFile.setSourcePath("driver/pic32cx-bz/src/src_bz3/app_fw/device_sleep.h")
     deviceSleepHeaderFile.setOutputName("device_sleep.h")
     deviceSleepHeaderFile.setDestPath('')
     deviceSleepHeaderFile.setProjectPath('config/' + configName)
@@ -585,7 +716,10 @@ def instantiateComponent(libBTZBCore):
     # Add device_sleep.c
     global deviceSleepSourceFile
     deviceSleepSourceFile = libBTZBCore.createFileSymbol("DEVICE_SLEEP_C", None)
-    deviceSleepSourceFile.setSourcePath("driver/pic32cx-bz/src/sleep/device_sleep.c")
+    if (deviceName in pic32cx_bz2_family):
+        deviceSleepSourceFile.setSourcePath("driver/pic32cx-bz/templates/low_power/device_sleep_bz2.ftl")
+    elif (deviceName in pic32cx_bz3_family):
+        deviceSleepSourceFile.setSourcePath("driver/pic32cx-bz/templates/low_power/device_sleep_bz3.ftl")
     deviceSleepSourceFile.setOutputName("device_sleep.c")
     deviceSleepSourceFile.setDestPath('')
     deviceSleepSourceFile.setProjectPath('config/' + configName)
@@ -594,6 +728,59 @@ def instantiateComponent(libBTZBCore):
     #setEnabled is controlled in handleRTC_Support
     deviceSleepSourceFile.setEnabled(False)
 
+    ############################################################################
+    ### Add ROM API Security Headers if BZ3 family is used
+    ############################################################################
+    if (deviceName in pic32cx_bz3_family):
+        rom_apiAPIHeaders = [
+            ('api_table.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/', 'driver/security/'),
+
+            ('adapter_api.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/silexpk/', 'driver/security/silexpk/'),
+            ('cmddefs_api.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/silexpk/', 'driversecurity/silexpk/'),
+            ('core_api.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/silexpk/', 'driver/security/silexpk/'),
+            ('dsa_api.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/silexpk/', 'driver/security/silexpk/'),
+            ('ec_curves_api.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/silexpk/', 'driver/security/silexpk/'),
+            ('eccweierstrass_api.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/silexpk/', 'driver/security/silexpk/'),
+            ('ecjpake.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/silexpk/', 'driver/security/silexpk/'),
+            ('ed448_api.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/silexpk/', 'driver/security/silexpk/'),
+            ('ed25519_api.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/silexpk/', 'driver/security/silexpk/'),
+            ('impl.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/silexpk/', 'driver/security/silexpk/'),
+            ('inputslots.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/silexpk/', 'driver/security/silexpk/'),
+            ('internal.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/silexpk/', 'driver/security/silexpk/'),
+            ('iomem.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/silexpk/', 'driver/security/silexpk/'),
+            ('montgomery_api.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/silexpk/', 'driver/security/silexpk/'),
+            ('rsa.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/silexpk/', 'driver/security/silexpk/'),
+            ('statuscodes_api.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/silexpk/', 'driver/security/silexpk/'),
+            ('sxbufop.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/silexpk/', 'driver/security/silexpk/'),
+            ('version.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/silexpk/', 'driver/security/silexpk/'),
+
+            ('aead_api.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/sxsymcrypt/', 'driver/security/sxsymcrypt/'),
+            ('blkcipher_api.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/sxsymcrypt/', 'driver/security/sxsymcrypt/'),
+            ('cmac_api.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/sxsymcrypt/', 'driver/security/sxsymcrypt/'),
+            ('cmmask_api.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/sxsymcrypt/', 'driver/security/sxsymcrypt/'),
+            ('hash_api.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/sxsymcrypt/', 'driver/security/sxsymcrypt/'),
+            ('hmac_api.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/sxsymcrypt/', 'driver/security/sxsymcrypt/'),
+            ('internal.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/sxsymcrypt/', 'driver/security/sxsymcrypt/'),
+            ('interrupts_api.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/sxsymcrypt/', 'driver/security/sxsymcrypt/'),
+            ('keyref_api.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/sxsymcrypt/', 'driver/security/sxsymcrypt/'),
+            ('mac_api.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/sxsymcrypt/', 'driver/security/sxsymcrypt/'),
+            ('sha1_api.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/sxsymcrypt/', 'driver/security/sxsymcrypt/'),
+            ('sha2_api.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/sxsymcrypt/', 'driver/security/sxsymcrypt/'),
+            ('sm3_api.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/sxsymcrypt/', 'driver/security/sxsymcrypt/'),
+            ('statuscodes.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/sxsymcrypt/', 'driver/security/sxsymcrypt/'),
+            ('trng_api.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/sxsymcrypt/', 'driver/security/sxsymcrypt/'),
+            ('trnginternal.h', 'driver/pic32cx-bz/src/src_bz3/rom_api/sxsymcrypt/', 'driver/security/sxsymcrypt/'),
+            ]
+
+        for incFile, srcPath,dstPath in rom_apiAPIHeaders:
+            incFileSym = libBTZBCore.createFileSymbol(None, None)
+            incFileSym.setSourcePath(srcPath + incFile)
+            incFileSym.setOutputName(incFile)
+            incFileSym.setOverwrite(True)
+            incFileSym.setDestPath(dstPath)
+            incFileSym.setProjectPath('config/' + configName + '/' + dstPath)
+            incFileSym.setType('HEADER')
+            incFileSym.setEnabled(True)
 
     ############################################################################
     ### Add logic for adding BLE stack to app.c
@@ -616,7 +803,24 @@ def instantiateComponent(libBTZBCore):
                                       'ZIGBEE_COLOR_LIGHT' or 'ZIGBEE_EXTENDED_COLOR_LIGHT' or 'ZIGBEE_TEMPERATURE_COLOR_LIGHT' or 
                                       'ZIGBEE_CUSTOM' or 'ZIGBEE_MACWSNTESTER' or 'ZIGBEE_GPD_SENSOR' or 'ZIGBEE_ZAPPSI'] in activeComponents)
     zigbeeStackLoaded.setVisible(False)
+    
+    ############################################################################
+    ### Add logic for adding Thread stack to app.c
+    ############################################################################
+    # This boolean is controlled configAppIdleTask called by:
+    #   onAttachmentConnected or onAttachmentDisconnected
+    threadStackLoaded = libBTZBCore.createBooleanSymbol('THREADSTACK_LOADED', None)
+    threadStackLoaded.setDefaultValue('OPEN_THREAD' in activeComponents)
+    threadStackLoaded.setVisible(False)
 
+    ############################################################################
+    ### Add logic for adding PHY to app.c
+    ############################################################################
+    # This boolean is controlled configAppIdleTask called by:
+    #   onAttachmentConnected or onAttachmentDisconnected
+    phyLoaded = libBTZBCore.createBooleanSymbol('802154_PHY_LOADED', None)
+    phyLoaded.setDefaultValue('IEEE_802154_PHY' in activeComponents)
+    phyLoaded.setVisible(False)
 
     ############################################################################
     ### Add logic for setting wssEnable_t mode
@@ -639,7 +843,7 @@ def instantiateComponent(libBTZBCore):
 
 
     ############################################################################
-    ### Add logic for adding Zigbee stack to app.c
+    ### Add logic for adding Zigbee/Thread stack to app.c
     ############################################################################
     global uartEnable
     uartEnable = libBTZBCore.createBooleanSymbol("ENABLE_CONSOLE", None)
@@ -648,7 +852,7 @@ def instantiateComponent(libBTZBCore):
     uartEnable.setVisible(False)
     uartEnable.setDependencies(isConsoleUartEnabled, ["ZIGBEE_COLOR_SCENE_CONTROLLER:APP_ENABLE_CONSOLE", "ZIGBEE_MULTI_SENSOR:APP_ENABLE_CONSOLE", "ZIGBEE_COMBINED_INTERFACE:APP_ENABLE_CONSOLE", "ZIGBEE_THERMOSTAT:APP_ENABLE_CONSOLE",
                                                    "ZIGBEE_IAS_ACE:APP_ENABLE_CONSOLE", "ZIGBEE_ON_OFF_LIGHT:APP_ENABLE_CONSOLE", "ZIGBEE_DIMMABLE_LIGHT:APP_ENABLE_CONSOLE", "ZIGBEE_COLOR_LIGHT:APP_ENABLE_CONSOLE",
-                                                   "ZIGBEE_EXTENDED_COLOR_LIGHT:APP_ENABLE_CONSOLE","ZIGBEE_TEMPERATURE_COLOR_LIGHT:APP_ENABLE_CONSOLE","ZIGBEE_CUSTOM:APP_ENABLE_CONSOLE","ZIGBEE_RUNNER:APP_ENABLE_CONSOLE"])
+                                                   "ZIGBEE_EXTENDED_COLOR_LIGHT:APP_ENABLE_CONSOLE","ZIGBEE_TEMPERATURE_COLOR_LIGHT:APP_ENABLE_CONSOLE","ZIGBEE_CUSTOM:APP_ENABLE_CONSOLE","ZIGBEE_RUNNER:APP_ENABLE_CONSOLE","OPEN_THREAD:OPEN_THREAD_UART_SERVICE"])
 
     ############################################################################
     ### Add idle task configuration items
@@ -681,17 +885,107 @@ def instantiateComponent(libBTZBCore):
     ############################################################################
     global deepSleepEnable
     deepSleepEnable = libBTZBCore.createBooleanSymbol("ENABLE_DEEP_SLEEP", None)
-    deepSleepEnable.setLabel("Enable Deep Sleep")
-    deepSleepEnable.setDefaultValue(False)
-    deepSleepEnable.setVisible(False)
+    deepSleepEnable.setLabel("Enable Deep Sleep (Retention RAM)")
+    deepSleepEnable.setDefaultValue(getBLEStackLibDsadven())
+    deepSleepEnable.setVisible(True)
+    deepSleepEnable.setDependencies(enableRetentionRAMSetting, ["BLE_STACK_LIB:GAP_DSADV_EN", "ENABLE_DEEP_SLEEP"])
+
+    ############################################################################
+    ### Add logic for retention RAM menu item - Stack Requirements
+    ############################################################################
+    global stackRetentionRAM
+    stackRetentionRAM = libBTZBCore.createIntegerSymbol('STACK_RETENTION_RAM', deepSleepEnable)
+    stackRetentionRAM.setLabel("Stack Retention RAM Size (bytes)")
+    stackRetentionRAM.setVisible(getDeepSleepState())
+    stackRetentionRAM.setDefaultValue(computeRetentionRAMRequirements())
+    stackRetentionRAM.setDependencies(enableStackRetentionRAM, ["ENABLE_DEEP_SLEEP"])
+
+    ############################################################################
+    ### Add logic for retention RAM menu item - App Requirements
+    ############################################################################
+    global appRetentionRAM
+    appRetentionRAM = libBTZBCore.createIntegerSymbol('APP_RETENTION_RAM', deepSleepEnable)
+    appRetentionRAM.setLabel("Application Retention RAM Size (bytes)")
+    appRetentionRAM.setVisible(getDeepSleepState())
+    appRetentionRAM.setDefaultValue(0)
+    appRetentionRAM.setMin(0) 
+    appRetentionRAM.setMax(computeMaxAppRetentionRAM())
+    appRetentionRAM.setDependencies(enableAppRetentionRAM, ["STACK_RETENTION_RAM"])
+
+    ############################################################################
+    ### Add logic for retention RAM option
+    ############################################################################
+    global totalRetentionRAM
+    totalRetentionRAM = libBTZBCore.createBooleanSymbol('TOTAL_RETENTION_RAM', deepSleepEnable)
+    totalRetentionRAM.setVisible(False)
+    totalRetentionRAM.setDefaultValue(False)
+    totalRetentionRAM.setDependencies(enable32KRetentionRAM, ["APP_RETENTION_RAM", "STACK_RETENTION_RAM"])
+
+    ############################################################################
+    ### Add logic for retention RAM option
+    ############################################################################
+    global totalRetentionRAMBytes
+    totalRetentionRAMBytes = libBTZBCore.createIntegerSymbol('TOTAL_RETENTION_RAM_BYTES', deepSleepEnable)
+    totalRetentionRAMBytes.setVisible(False)
+    totalRetentionRAMBytes.setDefaultValue(0)
 
     ############################################################################
     ### Add Custom Antenna Region
     ############################################################################
-    customAntennaRegion = libBTZBCore.createComboSymbol('CUSTOM_ANT_REGION', None, ['ETSI', 'FCC', 'ETSI_FCC'])
+    #customAntennaRegion = libBTZBCore.createComboSymbol('CUSTOM_ANT_REGION', None, ['ETSI', 'FCC', 'ETSI_FCC'])
+    customAntennaRegion = libBTZBCore.createMenuSymbol('CUSTOM_ANT_REGION', None)
     customAntennaRegion.setLabel('Regulatory Region')
     customAntennaRegion.setVisible(True)
+    
+    ############################################################################
+    ### Add logic for adding ResetToFN in device_deep_sleep.c
+    ############################################################################
+    global resetToFNEnable
+    resetToFNEnable = libBTZBCore.createBooleanSymbol("ENABLE_RESET_TO_FN", None)
+    resetToFNEnable.setDefaultValue(False)
+    resetToFNEnable.setVisible(False)
 
+    #ETSI/UK
+    etsiAntennaRegion = libBTZBCore.createBooleanSymbol("ETSI_REGION", customAntennaRegion)
+    etsiAntennaRegion.setLabel("ETSI/UK")
+    etsiAntennaRegion.setDefaultValue(True)
+    etsiAntennaRegion.setVisible(True)
+    etsiAntennaRegion.setReadOnly(False)
+
+    #FCC/IC
+    fccAntennaRegion = libBTZBCore.createBooleanSymbol("FCC_REGION", customAntennaRegion)
+    fccAntennaRegion.setLabel("FCC/IC")
+    fccAntennaRegion.setDefaultValue(False)
+    fccAntennaRegion.setVisible(True)
+    fccAntennaRegion.setReadOnly(False)
+
+    #JAPAN
+    japanAntennaRegion = libBTZBCore.createBooleanSymbol("JAPAN_REGION", customAntennaRegion)
+    japanAntennaRegion.setLabel("JAPAN")
+    japanAntennaRegion.setDefaultValue(False)
+    japanAntennaRegion.setVisible(True)
+    japanAntennaRegion.setReadOnly(False)
+
+    #KOREA
+    koreaAntennaRegion = libBTZBCore.createBooleanSymbol("KOREA_REGION", customAntennaRegion)
+    koreaAntennaRegion.setLabel("KOREA")
+    koreaAntennaRegion.setDefaultValue(False)
+    koreaAntennaRegion.setVisible(True)
+    koreaAntennaRegion.setReadOnly(False)
+
+    #CHINA
+    chinaAntennaRegion = libBTZBCore.createBooleanSymbol("CHINA_REGION", customAntennaRegion)
+    chinaAntennaRegion.setLabel("CHINA")
+    chinaAntennaRegion.setDefaultValue(False)
+    chinaAntennaRegion.setVisible(True)
+    chinaAntennaRegion.setReadOnly(False)
+
+    #TAIWAN
+    taiwanAntennaRegion = libBTZBCore.createBooleanSymbol("TAIWAN_REGION", customAntennaRegion)
+    taiwanAntennaRegion.setLabel("TAIWAN")
+    taiwanAntennaRegion.setDefaultValue(False)
+    taiwanAntennaRegion.setVisible(True)
+    taiwanAntennaRegion.setReadOnly(False)
 
     ############################################################################
     ### Add Custom Antenna Gain Enable
@@ -699,7 +993,7 @@ def instantiateComponent(libBTZBCore):
     customAntennaGainEnable = libBTZBCore.createBooleanSymbol('CUSTOM_ANT_ENABLE', None)
     customAntennaGainEnable.setLabel('Enable Custom Antenna Gain')
     customAntennaGainEnable.setValue(False)
-    customAntennaGainEnable.setDependencies(antGainChanged, ["CUSTOM_ANT_GAIN", "CUSTOM_ANT_REGION"])
+    customAntennaGainEnable.setDependencies(antGainChanged, ["CUSTOM_ANT_GAIN", "ETSI_REGION", "FCC_REGION", "JAPAN_REGION", "KOREA_REGION", "CHINA_REGION", "TAIWAN_REGION"])
 
 
     ############################################################################
@@ -707,7 +1001,11 @@ def instantiateComponent(libBTZBCore):
     ############################################################################
     customAntennaGain = libBTZBCore.createIntegerSymbol('CUSTOM_ANT_GAIN', customAntennaGainEnable)
     customAntennaGain.setLabel('Custom Antenna Gain (dBm)')
-    customAntennaGain.setDefaultValue(3)
+    if( deviceName == 'WBZ450' ):
+        customAntennaGain.setDefaultValue(5)
+    else:
+        customAntennaGain.setDefaultValue(3)
+
     customAntennaGain.setMin(-5)
     customAntennaGain.setMax(6)
     customAntennaGain.setVisible(False)
@@ -767,7 +1065,33 @@ def instantiateComponent(libBTZBCore):
     pmuModeSelection.setDefaultValue('PMU_MODE_BUCK_PWM')
     pmuModeSelection.setVisible((pmuModeSettingEnable.getValue() == True))
     pmuModeSelection.setDependencies(enablePMUModeSettingOption, ["SYSTEM_ENABLE_PMUMODE_SETTING"])
-    
+
+    ############################################################################
+    ### PIC32CX-BZ2 family Identification symbol
+    ############################################################################
+    # This boolean identifies the device is PIC32CX-BZ2 family device
+    pic32cxbzDevice = libBTZBCore.createBooleanSymbol('PIC32CX_BZ2_DEVICE', None)
+    pic32cxbzDevice.setDefaultValue(deviceName in pic32cx_bz2_family)
+    pic32cxbzDevice.setVisible(False)
+
+    # This boolean identifies the device is PIC32CX-BZ2 48 or 32 pin device
+    pic32cxbzPinDevice = libBTZBCore.createBooleanSymbol('PIC32CX_BZ2_48PIN_DEVICE', None)
+    pic32cxbzPinDevice.setDefaultValue(deviceName in pic32cx_bz2_48pin_family)
+    pic32cxbzPinDevice.setVisible(False)
+
+    ############################################################################
+    ### PIC32CX-BZ3 family Identification symbol
+    ############################################################################
+    # This boolean identifies the device is PIC32CX-BZ3 family device
+    pic32cxbznDevice = libBTZBCore.createBooleanSymbol('PIC32CX_BZ3_DEVICE', None)
+    pic32cxbznDevice.setDefaultValue(deviceName in pic32cx_bz3_family)
+    pic32cxbznDevice.setVisible(False)
+
+    # This boolean identifies the device is PIC32CX-BZ2 48 or 32 pin device
+    pic32cxbznPinDevice = libBTZBCore.createBooleanSymbol('PIC32CX_BZ3_48PIN_DEVICE', None)
+    pic32cxbznPinDevice.setDefaultValue(deviceName in pic32cx_bz3_48pin_family)
+    pic32cxbznPinDevice.setVisible(False)
+
 def finalizeComponent(libBTZBCore):
     print('Finalizing: libBTZBCore')
 
@@ -782,7 +1106,8 @@ def finalizeComponent(libBTZBCore):
         Database.setSymbolValue("FreeRTOS", "FREERTOS_USE_QUEUE_SETS", True)
 
     print("Enabling 'Generate Harmony Application Files'")
-    Database.sendMessage("HarmonyCore", "ENABLE_APP_FILE", {"isEnabled":True})
+    Database.sendMessage("HarmonyCore", "ENABLE_DRV_COMMON", {"isEnabled":True})
+    Database.sendMessage("HarmonyCore", "ENABLE_SYS_COMMON", {"isEnabled":True})
 
     print("Setting 'Minimal Stack Size' to 256")
     Database.setSymbolValue("FreeRTOS", "FREERTOS_MINIMAL_STACK_SIZE", 256)
@@ -794,7 +1119,11 @@ def finalizeComponent(libBTZBCore):
     if (remoteComponent):
         memoryManagementChoice = remoteComponent.getSymbolByID("FREERTOS_MEMORY_MANAGEMENT_CHOICE")
         memoryManagementChoice.setValue('Heap_4')
-
+        
+def destroyComponent(libBTZBCore):
+    Database.sendMessage("HarmonyCore", "ENABLE_DRV_COMMON", {"isEnabled":False})
+    Database.sendMessage("HarmonyCore", "ENABLE_SYS_COMMON", {"isEnabled":False})
+    
 def onAttachmentConnected(source, target):
     targetID = target["component"].getID()
     sourceID = source["component"].getID()
@@ -811,7 +1140,14 @@ def onAttachmentConnected(source, target):
         # Log.writeInfoMessage('device_support:onAttachmentConnected: CUSTOM_ANT_ENABLE symbol: {}'.format(source["component"].getSymbolByID("CUSTOM_ANT_ENABLE").getID()))
         sendAntMessage('CUSTOM_ANT_ENABLE', source["component"].getSymbolByID("CUSTOM_ANT_ENABLE").getValue())
         sendAntMessage('CUSTOM_ANT_GAIN', source["component"].getSymbolByID("CUSTOM_ANT_GAIN").getValue())
-        sendAntMessage('CUSTOM_ANT_REGION', source["component"].getSymbolByID("CUSTOM_ANT_REGION").getValue())
+        sendAntMessage('ETSI_REGION', source["component"].getSymbolByID("ETSI_REGION").getValue())
+        sendAntMessage('FCC_REGION', source["component"].getSymbolByID("FCC_REGION").getValue())
+        sendAntMessage('JAPAN_REGION', source["component"].getSymbolByID("JAPAN_REGION").getValue())
+        sendAntMessage('KOREA_REGION', source["component"].getSymbolByID("KOREA_REGION").getValue())
+        sendAntMessage('CHINA_REGION', source["component"].getSymbolByID("CHINA_REGION").getValue())
+        sendAntMessage('TAIWAN_REGION', source["component"].getSymbolByID("TAIWAN_REGION").getValue())
+
+
 
     if targetID == 'rtc':
         Database.setSymbolValue("rtc", "RTC_MODE0_INTENSET_CMP0_ENABLE", True)
@@ -859,7 +1195,7 @@ def handleRTC_Support(args):
     localComponent = Database.getComponentByID(args['target'])
     if (localComponent):
         localComponentID = localComponent.getID()
-        rtcRequiredSymbol = localComponent.getSymbolByID('SLEEP_SUPPORTED')
+        rtcRequiredSymbol = (localComponent.getSymbolByID('SLEEP_SUPPORTED') or localComponent.getSymbolByID('ENABLE_DEEP_SLEEP'))
 
     Log.writeInfoMessage('device_support:handleRTC_Support target ID = {}'.format(localComponentID))
 
@@ -886,6 +1222,7 @@ def handleRTC_Support(args):
 # Dependency callback called upon PMU Mode Setting is enabled
 def enablePMUModeSettingOption(symbol, event):
     symbol.setVisible(event["value"])
+
 def handleDeepSleepEnable(args):
     deepSleepEnable.setValue(args["isEnabled"])
     global deviceDeepSleepHeaderFile
@@ -893,11 +1230,70 @@ def handleDeepSleepEnable(args):
     if (args["isEnabled"] == True):
         deviceDeepSleepHeaderFile.setEnabled(True)
         deviceDeepSleepSourceFile.setEnabled(True)
+        deepSleepEnable.setReadOnly(True)
         print("handleDeepSleepEnable setting True")
     else:
         deviceDeepSleepHeaderFile.setEnabled(False)
         deviceDeepSleepSourceFile.setEnabled(False)
-        print("handleDeepSleepEnable setting False")    
+        deepSleepEnable.setReadOnly(False)
+        print("handleDeepSleepEnable setting False")
+
+def enableRetentionRAMSetting(symbol, event):
+    if ((event["value"] == True)):
+        deepSleepEnable.setValue(True)
+        if(event["source"] == "BLE_STACK_LIB:GAP_DSAVD_EN"):
+            print("GAP_DSADV_EN enabled")
+        else:
+            print("enable deep sleep setting enabled")
+            if (deviceName in pic32cx_bz2_family):
+                handleRTC_Support({"target": "pic32cx_bz2_devsupport", "source": "pic32cx_bz2_devsupport", "rtcRequired": True})
+                deviceDeepSleepHeaderFile.setEnabled(True)
+                deviceDeepSleepSourceFile.setEnabled(True)
+            elif (deviceName in pic32cx_bz3_family):
+                handleRTC_Support({"target": "pic32cx_bz3_devsupport", "source": "pic32cx_bz3_devsupport", "rtcRequired": True})
+                deviceDeepSleepHeaderFile.setEnabled(True)
+                deviceDeepSleepSourceFile.setEnabled(True)
+    else:
+        deepSleepEnable.setValue(False)
+        if(event["source"] == "BLE_STACK_LIB:GAP_DSAVD_EN"):
+            print("GAP_DSADV_EN disabled")
+        else:
+            print("enable deep sleep setting disabled")
+            if (deviceName in pic32cx_bz2_family):
+                handleRTC_Support({"target": "pic32cx_bz2_devsupport", "source": "pic32cx_bz2_devsupport", "rtcRequired": False})
+                deviceDeepSleepHeaderFile.setEnabled(False)
+                deviceDeepSleepSourceFile.setEnabled(False)
+            elif (deviceName in pic32cx_bz3_family):
+                handleRTC_Support({"target": "pic32cx_bz3_devsupport", "source": "pic32cx_bz3_devsupport", "rtcRequired": False})
+                deviceDeepSleepHeaderFile.setEnabled(False)
+                deviceDeepSleepSourceFile.setEnabled(False)
+
+def enableStackRetentionRAM(symbol, event):
+    if(getDeepSleepState() and (deviceName in pic32cx_bz3_family)):
+        stackRetentionRAM.setValue(computeRetentionRAMRequirements())
+        stackRetentionRAM.setVisible(True)
+        stackRetentionRAM.setReadOnly(True)
+    else:
+        stackRetentionRAM.setVisible(False)
+        stackRetentionRAM.setReadOnly(False)
+
+def enableAppRetentionRAM(symbol, event):
+    if(getDeepSleepState() and (deviceName in pic32cx_bz3_family)):
+        appRetentionRAM.setMax(computeMaxAppRetentionRAM())
+        appRetentionRAM.setVisible(True)
+    else:
+        appRetentionRAM.setVisible(False)
+
+def enable32KRetentionRAM(symbol, event):
+    if(deviceName in pic32cx_bz3_family):
+        appvalue = int(Database.getSymbolValue("pic32cx_bz3_devsupport", "APP_RETENTION_RAM"))
+        configureRetentionRAM = int(appvalue + int(stackRetentionRAM.getValue()))
+        totalRetentionRAMBytes.setValue(configureRetentionRAM)
+        if((configureRetentionRAM > 16384) and (configureRetentionRAM < 32768)):
+            totalRetentionRAM.setValue(True)
+        else:
+            totalRetentionRAM.setValue(False)
+
 
 def handleMessage(messageID, args):
     '''
@@ -922,3 +1318,4 @@ def handleMessage(messageID, args):
         handleRTC_Support(args)
     elif (messageID == "DEEP_SLEEP_ENABLE"):
         handleDeepSleepEnable(args)
+